+++ components/esp_lvgl_port/src/lvgl8/esp_lvgl_port_disp.c
@@ -8,6 +8,7 @@
 #include "esp_err.h"
 #include "esp_check.h"
 #include "esp_heap_caps.h"
+#include "esp_idf_version.h"
 #include "esp_lcd_panel_io.h"
 #include "esp_lcd_panel_ops.h"
 #include "esp_lvgl_port.h"