+++ components/esp_lvgl_port/src/lvgl9/esp_lvgl_port.c
@@ -25,7 +25,8 @@ static const char *TAG = "LVGL";
 
 typedef struct lvgl_port_ctx_s {
     SemaphoreHandle_t   lvgl_mux;
-    SemaphoreHandle_t   task_mux;
+    QueueHandle_t       lvgl_queue;
+    SemaphoreHandle_t   task_init_mux;
     esp_timer_handle_t  tick_timer;
     bool                running;
     int                 task_max_sleep_ms;
@@ -69,9 +70,12 @@ esp_err_t lvgl_port_init(const lvgl_port_cfg_t *cfg)
     /* LVGL semaphore */
     lvgl_port_ctx.lvgl_mux = xSemaphoreCreateRecursiveMutex();
     ESP_GOTO_ON_FALSE(lvgl_port_ctx.lvgl_mux, ESP_ERR_NO_MEM, err, TAG, "Create LVGL mutex fail!");
-    /* Task semaphore */
-    lvgl_port_ctx.task_mux = xSemaphoreCreateMutex();
-    ESP_GOTO_ON_FALSE(lvgl_port_ctx.task_mux, ESP_ERR_NO_MEM, err, TAG, "Create LVGL task sem fail!");
+    /* Task init semaphore */
+    lvgl_port_ctx.task_init_mux = xSemaphoreCreateMutex();
+    ESP_GOTO_ON_FALSE(lvgl_port_ctx.task_init_mux, ESP_ERR_NO_MEM, err, TAG, "Create LVGL task sem fail!");
+    /* Task queue */
+    lvgl_port_ctx.lvgl_queue = xQueueCreate(10, sizeof(lvgl_port_event_t));
+    ESP_GOTO_ON_FALSE(lvgl_port_ctx.lvgl_queue, ESP_ERR_NO_MEM, err, TAG, "Create LVGL queue fail!");
 
     BaseType_t res;
     if (cfg->task_affinity < 0) {
@@ -128,7 +132,7 @@ esp_err_t lvgl_port_deinit(void)
     }
 
     /* Wait for stop task */
-    if (xSemaphoreTake(lvgl_port_ctx.task_mux, pdMS_TO_TICKS(ESP_LVGL_PORT_TASK_MUX_DELAY_MS)) != pdTRUE) {
+    if (xSemaphoreTake(lvgl_port_ctx.task_init_mux, pdMS_TO_TICKS(ESP_LVGL_PORT_TASK_MUX_DELAY_MS)) != pdTRUE) {
         ESP_LOGE(TAG, "Failed to stop LVGL task");
         return ESP_ERR_TIMEOUT;
     }
@@ -153,16 +157,36 @@ void lvgl_port_unlock(void)
     xSemaphoreGiveRecursive(lvgl_port_ctx.lvgl_mux);
 }
 
+esp_err_t lvgl_port_task_wake(lvgl_port_event_t event, bool isr)
+{
+    if (!lvgl_port_ctx.lvgl_queue) {
+        return ESP_ERR_INVALID_STATE;
+    }
+
+    if (isr) {
+        BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+        xQueueSendFromISR(lvgl_port_ctx.lvgl_queue, &event, &xHigherPriorityTaskWoken);
+        if (xHigherPriorityTaskWoken) {
+            portYIELD_FROM_ISR( );
+        }
+    } else {
+        xQueueSend(lvgl_port_ctx.lvgl_queue, &event, 0);
+    }
+
+    return ESP_OK;
+}
+
 /*******************************************************************************
 * Private functions
 *******************************************************************************/
 
 static void lvgl_port_task(void *arg)
 {
-    uint32_t task_delay_ms = lvgl_port_ctx.task_max_sleep_ms;
+    lvgl_port_event_t event = 0;
+    uint32_t task_delay_ms = 0;
 
     /* Take the task semaphore */
-    if (xSemaphoreTake(lvgl_port_ctx.task_mux, 0) != pdTRUE) {
+    if (xSemaphoreTake(lvgl_port_ctx.task_init_mux, 0) != pdTRUE) {
         ESP_LOGE(TAG, "Failed to take LVGL task sem");
         lvgl_port_task_deinit();
         vTaskDelete( NULL );
@@ -171,6 +195,9 @@ static void lvgl_port_task(void *arg)
     ESP_LOGI(TAG, "Starting LVGL task");
     lvgl_port_ctx.running = true;
     while (lvgl_port_ctx.running) {
+        /* Wait for queue or timeout (sleep task) */
+        xQueueReceive(lvgl_port_ctx.lvgl_queue, &event, pdMS_TO_TICKS(task_delay_ms));
+
         if (lv_display_get_default() && lvgl_port_lock(0)) {
             task_delay_ms = lv_timer_handler();
             lvgl_port_unlock();
@@ -180,11 +207,15 @@ static void lvgl_port_task(void *arg)
         } else if (task_delay_ms < 1) {
             task_delay_ms = 1;
         }
-        vTaskDelay(pdMS_TO_TICKS(task_delay_ms));
+
+        /* Sleep task, when everything done */
+        if (lv_anim_count_running() == 0 && lv_display_get_inactive_time(NULL) > 2 * LV_DEF_REFR_PERIOD) {
+            task_delay_ms = lvgl_port_ctx.task_max_sleep_ms;
+        }
     }
 
     /* Give semaphore back */
-    xSemaphoreGive(lvgl_port_ctx.task_mux);
+    xSemaphoreGive(lvgl_port_ctx.task_init_mux);
 
     /* Close task */
     vTaskDelete( NULL );
@@ -195,8 +226,11 @@ static void lvgl_port_task_deinit(void)
     if (lvgl_port_ctx.lvgl_mux) {
         vSemaphoreDelete(lvgl_port_ctx.lvgl_mux);
     }
-    if (lvgl_port_ctx.task_mux) {
-        vSemaphoreDelete(lvgl_port_ctx.task_mux);
+    if (lvgl_port_ctx.task_init_mux) {
+        vSemaphoreDelete(lvgl_port_ctx.task_init_mux);
+    }
+    if (lvgl_port_ctx.lvgl_queue) {
+        vQueueDelete(lvgl_port_ctx.lvgl_queue);
     }
     memset(&lvgl_port_ctx, 0, sizeof(lvgl_port_ctx));
 #if LV_ENABLE_GC || !LV_MEM_CUSTOM