+++ components/esp_lvgl_port/src/lvgl9/esp_lvgl_port_touch.c
@@ -26,6 +26,7 @@ typedef struct {
 *******************************************************************************/
 
 static void lvgl_port_touchpad_read(lv_indev_t *indev_drv, lv_indev_data_t *data);
+static void lvgl_port_touch_interrupt_callback(esp_lcd_touch_handle_t tp);
 
 /*******************************************************************************
 * Public API functions
@@ -33,7 +34,8 @@ static void lvgl_port_touchpad_read(lv_indev_t *indev_drv, lv_indev_data_t *data
 
 lv_indev_t *lvgl_port_add_touch(const lvgl_port_touch_cfg_t *touch_cfg)
 {
-    lv_indev_t *indev;
+    esp_err_t ret = ESP_OK;
+    lv_indev_t *indev = NULL;
     assert(touch_cfg != NULL);
     assert(touch_cfg->disp != NULL);
     assert(touch_cfg->handle != NULL);
@@ -46,6 +48,12 @@ lv_indev_t *lvgl_port_add_touch(const lvgl_port_touch_cfg_t *touch_cfg)
     }
     touch_ctx->handle = touch_cfg->handle;
 
+    if (touch_ctx->handle->config.int_gpio_num != GPIO_NUM_NC) {
+        /* Register touch interrupt callback */
+        ret = esp_lcd_touch_register_interrupt_callback(touch_ctx->handle, lvgl_port_touch_interrupt_callback);
+        ESP_GOTO_ON_ERROR(ret, err, TAG, "Error in register touch interrupt.");
+    }
+
     lvgl_port_lock(0);
     /* Register a touchpad input device */
     indev = lv_indev_create();
@@ -55,6 +63,13 @@ lv_indev_t *lvgl_port_add_touch(const lvgl_port_touch_cfg_t *touch_cfg)
     lv_indev_set_user_data(indev, touch_ctx);
     touch_ctx->indev = indev;
     lvgl_port_unlock();
+
+err:
+    if (ret != ESP_OK) {
+        if (touch_ctx) {
+            free(touch_ctx);
+        }
+    }
 
     return indev;
 }
@@ -68,6 +83,11 @@ esp_err_t lvgl_port_remove_touch(lv_indev_t *touch)
     /* Remove input device driver */
     lv_indev_delete(touch);
     lvgl_port_unlock();
+
+    if (touch_ctx->handle->config.int_gpio_num != GPIO_NUM_NC) {
+        /* Unregister touch interrupt callback */
+        esp_lcd_touch_register_interrupt_callback(touch_ctx->handle, NULL);
+    }
 
     if (touch_ctx) {
         free(touch_ctx);
@@ -104,4 +124,10 @@ static void lvgl_port_touchpad_read(lv_indev_t *indev_drv, lv_indev_data_t *data
     } else {
         data->state = LV_INDEV_STATE_RELEASED;
     }
+}
+
+static void lvgl_port_touch_interrupt_callback(esp_lcd_touch_handle_t tp)
+{
+    /* Wake LVGL task, if needed */
+    lvgl_port_task_wake(LVGL_PORT_EVENT_TOUCH, true);
 }