+++ components/esp_lvgl_port/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.0.1
+
+### Features
+- Added unit test
+- Allow to sleep main LVGL task
+- Wake LVGL task from touch and display, when set big maximum sleep
+
+### Fixes
+- Apply display rotation from configuration
+- Wait for stop task when deinit
+- Added esp_idf_version.h
+
 ## 2.0.0
 
 ### Features